%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  999.package.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{aroma.agilent-package}
\alias{aroma.agilent-package}
\alias{aroma.agilent}
\docType{package}

\title{Package aroma.agilent}


 \description{
   This package provides method for analysis of Agilent data.
 }

 \section{Installation and updates}{

   To install this package, see instructions at
   \url{http://www.aroma-project.org/}.
 }

 \author{
   Henrik Bengtsson.
 }

 \section{License}{
   The releases of this package is licensed under
   LGPL version 2.1 or newer.

   The development code of the packages is under a private licence
   (where applicable) and patches sent to the author fall under the
   latter license, but will be, if incorporated, released under the
   "release" license above.
 }

 \references{
  Please see \url{http://www.aroma-project.org/} for references.
 }
\keyword{package}
