%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  999.NonDocumentedObjects.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{Non-documented objects}
\alias{Non-documented objects}
\title{Non-documented objects}


% Other missing docs
\alias{AgilentDataFile}
\alias{as.character.AgilentDataFile}
\alias{countLines.AgilentDataFile}
\alias{getBarcode}
\alias{getBarcode.AgilentDataFile}
\alias{getColumnNames.AgilentDataFile}
\alias{getDimension}
\alias{getDimension.AgilentDataFile}
\alias{getHeader.AgilentDataFile}
\alias{getScanDate}
\alias{getScanDate.AgilentDataFile}
\alias{hasColumnHeader.AgilentDataFile}
\alias{readColumnFast}
\alias{readColumnFast.AgilentDataFile}
\alias{readColumnNames.AgilentDataFile}
\alias{readColumnsFast}
\alias{readColumnsFast.AgilentDataFile}
\alias{readFeatures}
\alias{readFeatures.AgilentDataFile}
\alias{readFeaturesColumn}
\alias{readFeaturesColumn.AgilentDataFile}
\alias{readLines.AgilentDataFile}
\alias{readRawHeader.AgilentDataFile}
\alias{readSection}
\alias{readSection.AgilentDataFile}
\alias{readSections}
\alias{readSections.AgilentDataFile}
\alias{readSectionsIdxs}
\alias{readSectionsIdxs.AgilentDataFile}
\alias{verify.AgilentDataFile}
\alias{readFeaturesColumns}
\alias{readFeaturesColumns.AgilentDataFile}
\alias{AgilentDataSet}
\alias{byPath.AgilentDataSet}
\alias{exportCopyNumbers}
\alias{exportCopyNumbers.AgilentDataFile}
\alias{getChipType.AgilentDataFile}
\alias{getChipType.AgilentDataSet}
\alias{getExtensionPattern.AgilentDataFile}
\alias{getPlatform.AgilentDataFile}
\alias{getPlatform.AgilentDataSet}
\alias{nbrOfLines.AgilentDataFile}
\alias{nbrOfRows.AgilentDataFile}

\description{
  This page contains aliases for all "non-documented" objects that
  \code{R CMD check} detects in this package.

  Almost all of them are \emph{generic} functions that have specific
  document for the corresponding method coupled to a specific class.
  Other functions are re-defined by \code{setMethodS3()} to
  \emph{default} methods. Neither of these two classes are non-documented
  in reality.
  The rest are deprecated methods.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}


\keyword{documentation}
\keyword{internal}
